Some backwards compatibility breaking changes, e.g. moving the search model classes to a different package.
