+++ src/main/java/com/sumologic/client/search/SearchClient.java
@@ -3,9 +3,9 @@ package com.sumologic.client.search;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.sumologic.client.Credentials;
 import com.sumologic.client.UrlParameters;
-import com.sumologic.client.model.LogMessage;
-import com.sumologic.client.model.SearchRequest;
-import com.sumologic.client.model.SearchResponse;
+import com.sumologic.client.search.model.LogMessage;
+import com.sumologic.client.search.model.SearchRequest;
+import com.sumologic.client.search.model.SearchResponse;
 import com.sumologic.client.util.HttpUtils;
 import com.sumologic.client.util.JacksonUtils;
 import com.sumologic.client.util.ResponseHandler;
